Add new mode output, and add icon

+++ Calc/src/calculator/Model.java
@@ -2,7 +2,7 @@ package calculator;
 
 public class Model {
 
-	public long calculate(long number1, long number2, String operator){
+	public double calculate(double number1, double number2, String operator){
 		switch(operator){
 			case "+":
 				return number1 + number2;
